Mark assembly code files as not requiring executable stack
Because of the use of assembly code for callNative.S, gcc marks
the resulting object files as having executable stack. This leads
to libjvm.so being marked as such, and problems ensue when using
the library on hardened kernels.

Originally submitted upstream to the JamVM mailing list [1], but
never merged. Carried by IcedTea as PR2500 (2.6) / PR2523 (3.x).

 [1]: https://sourceforge.net/p/jamvm/mailman/message/34278321/

[grodriguez: extended to AArch64]
Signed-off-by: Guillermo Rodríguez <grodriguez@ingelabs.com>