Add attributes for condvars used for relative timed waits
Condvars used with pthread_cond_timedwait() and involving
relative timeouts are now initialised with attributes from a
shared pthread_condattr_t object. These waits will later be
switched to a monotonic clock (if available).

In the case of park_cv, which can be used with either relative
or absolute timeouts, the existing condvar is split in two
(park_rel_cv / park_abs_cv). Also, when a thread is parked, it
is necessary to keep track of which condvar is in use; this is
done with a new park_abs flag.

Signed-off-by: Guillermo Rodriguez <grodriguez@ingelabs.com>